+++ packages/golden-layout/changelog.md
@@ -1,4 +1,5 @@
 2.3.1
+fix: added safety checks for minimizing items which are not maximized
 fix: improved the performance of the destroy function
 2.3.0
 feat: added support for maximization boundaries